+++ phydrax/discretization/flip/__init__.py
@@ -1,0 +1,36 @@
+#
+# Copyright © 2026 PHYDRA, Inc. All rights reserved.
+#
+
+"""Fixed-population free-surface FLIP over prepared MAC grids."""
+
+from ._method import FLIPMethodPlan, FLIPResourcePolicy
+from ._transfer import FLIPParticleTransferPlan, PreparedFLIPParticleTransfer
+from ._types import (
+    FLIPDiagnostics,
+    FLIPGridToParticleResult,
+    FLIPParticleState,
+    FLIPParticleToGridResult,
+    FLIPRejectionReason,
+    FLIPRunStatus,
+    FLIPRuntimeState,
+    FLIPStepResult,
+    FLIPTransferState,
+)
+
+
+__all__ = [
+    "FLIPDiagnostics",
+    "FLIPGridToParticleResult",
+    "FLIPMethodPlan",
+    "FLIPParticleState",
+    "FLIPParticleToGridResult",
+    "FLIPParticleTransferPlan",
+    "FLIPRejectionReason",
+    "FLIPResourcePolicy",
+    "FLIPRunStatus",
+    "FLIPRuntimeState",
+    "FLIPStepResult",
+    "FLIPTransferState",
+    "PreparedFLIPParticleTransfer",
+]